requirements.txt
streamlit
google-generativeai
arxiv
chromadb
sentence-transformers
PyMuPDF
python-dotenv
celery
redis
openai
langchain
pandas
numpy
jupyter
flask
fastapi
pytest
